src/main/webapp/app/entities/main/detail/main-detail.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { DataUtils } from 'app/core/util/data-util.service';

import { MainDetailComponent } from './main-detail.component';

describe('Component Tests', () => {
  describe('Main Management Detail Component', () => {
    let comp: MainDetailComponent;
    let fixture: ComponentFixture<MainDetailComponent>;
    let dataUtils: DataUtils;

    beforeEach(() => {
      TestBed.configureTestingModule({
        declarations: [MainDetailComponent],
        providers: [
          {
            provide: ActivatedRoute,
            useValue: { data: of({ main: { id: 123 } }) },
          },
        ],
      })
        .overrideTemplate(MainDetailComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(MainDetailComponent);
      comp = fixture.componentInstance;
      dataUtils = TestBed.inject(DataUtils);
    });

    describe('OnInit', () => {
      it('Should load main on init', () => {
        // WHEN
        comp.ngOnInit();

        // THEN
        expect(comp.main).toEqual(jasmine.objectContaining({ id: 123 }));
      });
    });

    describe('byteSize', () => {
      it('Should call byteSize from DataUtils', () => {
        // GIVEN
        spyOn(dataUtils, 'byteSize');
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.byteSize(fakeBase64);

        // THEN
        expect(dataUtils.byteSize).toBeCalledWith(fakeBase64);
      });
    });

    describe('openFile', () => {
      it('Should call openFile from DataUtils', () => {
        // GIVEN
        spyOn(dataUtils, 'openFile');
        const fakeContentType = 'fake content type';
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.openFile(fakeBase64, fakeContentType);

        // THEN
        expect(dataUtils.openFile).toBeCalledWith(fakeBase64, fakeContentType);
      });
    });
  });
});
